backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.100.0
anyio==4.13.0
certifi==2026.4.22
click==8.3.3
colorama==0.4.6
distro==1.9.0
docstring_parser==0.18.0
ecdsa==0.19.2
fastapi==0.136.1
greenlet==3.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.14
jiter==0.14.0
Mako==1.3.12
MarkupSafe==3.0.3
passlib==1.7.4
psycopg2-binary==2.9.12
pyasn1==0.6.3
pydantic==2.13.4
pydantic-settings==2.14.1
pydantic_core==2.46.4
python-dotenv==1.2.2
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
